ensures document is saved when adding annotation

+++ peachjam/js/components/Annotations/Annotation.vue
@@ -1,6 +1,6 @@
 <template>
   <la-gutter-item
-    :id="annotation.id"
+    :id="annotation.ref_id"
     :anchor.prop="anchorElement"
   >
     <div class="card">
@@ -23,31 +23,34 @@
             </div>
           </div>
         </div>
-        <textarea
-          v-if="editing"
-          v-model="annotation.text"
-          class="form-control grow-text"
-        />
-        <p v-else>
-          {{ annotation.text }}
-        </p>
-        <div class="mt-2 text-end">
-          <button
+        <form @submit="saveAnnotation">
+          <textarea
             v-if="editing"
-            class="ms-1 btn btn-sm btn-secondary"
-            @click="editing = false"
-          >
-            Cancel
-          </button>
+            v-model="annotation.text"
+            class="form-control grow-text"
+            required
+          />
+          <p v-else>
+            {{ annotation.text }}
+          </p>
+          <div class="mt-2 text-end">
+            <button
+              v-if="editing"
+              class="ms-1 btn btn-sm btn-secondary"
+              @click="cancelEdit"
+            >
+              Cancel
+            </button>
 
-          <button
-            v-if="editing"
-            class="ms-1 btn btn-sm btn-primary"
-            @click="saveAnnotation"
-          >
-            Save
-          </button>
-        </div>
+            <button
+              v-if="editing"
+              class="ms-1 btn btn-sm btn-primary"
+              type="submit"
+            >
+              Save
+            </button>
+          </div>
+        </form>
       </div>
     </div>
   </la-gutter-item>
@@ -65,22 +68,42 @@ export default {
       default: null
     },
     viewRoot: HTMLElement,
-    gutter: HTMLElement,
-    useSelectors: Boolean
+    gutter: HTMLElement
   },
+  emits: ['remove-annotation'],
   data: (x) => ({
     marks: [],
     anchorElement: null,
     editing: x.annotationData.editing || false,
-    annotation: x.annotationData
+    annotation: x.annotationData,
+    ref_id: x.annotationData.ref_id
   }),
   mounted () {
     this.anchorElement = document.getElementById(this.annotation.target_id);
     this.mark();
     this.gutter.appendChild(this.$el);
   },
+  unmounted () {
+    this.unmark();
+  },
+
   methods: {
-    async saveAnnotation () {
+    async checkDocumentSaved () {
+      const response = await fetch(`/saved-documents/button/${this.viewRoot.dataset.documentId}`);
+      if (!response.ok) {
+        throw new Error(`Error checkin document saved: ${response.status}`);
+      }
+      return response.url.includes('update');
+    },
+    async saveAnnotation (e) {
+      e.preventDefault();
+      const isSaved = await this.checkDocumentSaved();
+      if (!isSaved) {
+        await window.htmx.ajax('post', `/saved-documents/create?doc_id=${this.annotation.document}`, {
+          target: '.save-document-button'
+        });
+      }
+
       const headers = await authHeaders();
       headers['Content-Type'] = 'application/json';
 
@@ -91,8 +114,8 @@ export default {
         document: this.annotation.document
       };
 
-      const method = this.annotation.id === 'new' ? 'POST' : 'PUT';
-      const url = this.annotation.id === 'new' ? `/api/documents/${this.annotation.document}/annotations/` : `/api/documents/${this.annotation.document}/annotations/${this.annotation.id}/`;
+      const method = this.annotation.ref_id.includes('new') ? 'POST' : 'PUT';
+      const url = this.annotation.ref_id.includes('new') ? `/api/documents/${this.annotation.document}/annotations/` : `/api/documents/${this.annotation.document}/annotations/${this.annotation.id}/`;
       const resp = await fetch(url, {
         method,
         headers,
@@ -101,6 +124,7 @@ export default {
 
       if (resp.ok) {
         this.annotation = await resp.json();
+        this.annotation.ref_id = this.ref_id;
         this.editing = false;
       }
     },
@@ -108,15 +132,17 @@ export default {
       this.editing = true;
     },
     async deleteAnnotation () {
-      const headers = await authHeaders();
-      const resp = await fetch(`/api/documents/${this.annotation.document}/annotations/${this.annotation.id}/`, {
-        method: 'DELETE',
-        headers
-      });
-      if (resp.ok) {
-        this.unmark();
-        this.$el.remove();
+      if (this.annotation.id) {
+        const headers = await authHeaders();
+        const resp = await fetch(`/api/documents/${this.annotation.document}/annotations/${this.annotation.id}/`, {
+          method: 'DELETE',
+          headers
+        });
+        if (!resp.ok) {
+          throw new Error('Failed to delete annotation');
+        }
       }
+      this.$emit('remove-annotation', this.annotation);
     },
     mark () {
       const range = targetToRange({
@@ -129,6 +155,12 @@ export default {
         return mark;
       }
       );
+    },
+    cancelEdit () {
+      this.editing = false;
+      if (this.annotation.ref_id.includes('new') && !this.annotation.id) {
+        this.$emit('remove-annotation', this.annotation);
+      }
     },
     unmark () {
       this.marks.forEach(mark => {
